Refactored 'align_images_using_neighbors' into its own file, given its sheer size
